+++ client/src/modules/folder/FolderDetails.tsx
@@ -3,9 +3,8 @@ import { NextPage } from 'next';
 import { useRouter } from 'next/router';
 import { FolderForm } from '@src/modules/folder/FolderForm';
 import { FolderDelete } from '@src/modules/folder/FolderDelete';
-import { ActionList } from '@src/ui/ActionList';
+import { ActionList, DropdownMenu } from '@src/ui';
 import { MoreIcon } from '@src/ui/Icons';
-import { DropdownMenu } from '@src/ui';
 import { FolderInterface } from '@src/interfaces';
 import { setApi } from '@src/apis';
 import { useQuery } from 'react-query';
