+++ client/src/ui/Spinner.tsx
@@ -1,4 +1,4 @@
-export function Spinner({ className, ...rest }: { className?: string }) {
+function Spinner({ className, ...rest }: { className?: string }) {
   return (
     <svg
       xmlns="http://www.w3.org/2000/svg"
@@ -64,3 +64,5 @@ export function Spinner({ className, ...rest }: { className?: string }) {
     </svg>
   );
 }
+
+export default Spinner;
