+++ client/src/ui/index.ts
@@ -1,4 +1,6 @@
+export { default as Layout } from './Layout';
 export { default as DropdownMenu } from './DropdownMenu';
 export { default as Dialog } from './Dialog';
 export { default as Textarea } from './Textarea';
-export { ActionList } from './ActionList';
+export { default as ActionList } from './ActionList';
+export { default as Spinner } from './Spinner';
